iteration1/requirements.txt
python-dotenv
openai==0.28
pandas
plotly
streamlit
openpyxl
kaleido